Rename Entry component to Content
Render navigation as part of entry.

REDMINE-
