+++ entry_types/scrolled/package/spec/support/pageObjects.js
@@ -1,7 +1,7 @@
 import React from 'react';
 
 import {renderInEntry} from './rendering';
-import Entry from 'frontend/Entry';
+import {Entry} from 'frontend/Entry';
 import foregroundStyles from 'frontend/Foreground.module.css';
 import {loadInlineEditingComponents} from 'frontend/inlineEditing';
 import {api} from 'frontend/api';
